move rdelete command to archiver.rdelete
